Added -ve lookup from readdir cache.
We add a notion of "confirmed". If a dir cache is confirmed it means
we know for sure that a particular file which was not found in the
dir cache is definitely not present and we can tell fuse about that.
All newly created directories start as "confirmed". Also a directory
cache becomes confirmed if it's full enumerated (from cookie=0 to eof).